change the sort of skillset by name, create page.py page.
